move control-flow-unflattening file into dynamics folder
